Partie 2 fonctionnelle

+++ debug/run.sh
@@ -1,5 +1,12 @@
-./run-qemu.sh &
-pidQemu=$!
-./run-gdb.sh
-kill $pidQemu
-killall qemu-system-arm
+cd ../
+if make
+then
+	cd debug/
+	./run-qemu.sh &
+	pidQemu=$!
+	./run-gdb.sh
+	kill $pidQemu
+	killall qemu-system-arm
+else
+	print 'Compilation fault';
+fi
